fix(scripts/build/setup/termux_setup_python_pip): only copy `sysconfig_module` if it does not already exist in the destination
- Fixes `cp: '/home/builder/.termux-build/python3.12-crossenv-prefix-bionic-i686/build/lib/python3.12/sysconfig.py' and '/home/builder/.termux-build/python3.12-crossenv-prefix-bionic-i686/build/lib/python3.12/sysconfig.py' are the same file`

+++ scripts/build/setup/termux_setup_python_pip.sh
@@ -70,7 +70,9 @@ termux_setup_python_pip() {
 		export PATH="${_CROSS_PATH}:$(echo -n $(tr ':' '\n' <<< "${PATH}" | grep -v "^${_CROSS_PATH}$") | tr ' ' ':')"
 
 		local sysconfig_module=$(${TERMUX_PYTHON_CROSSENV_PREFIX}/build/bin/python -c "import sysconfig; print(sysconfig.__file__)")
-		cp -r "${sysconfig_module}" "${TERMUX_PYTHON_CROSSENV_BUILDHOME}"
-		sed -i "s|os.path.normpath(sys.*prefix)|\"${TERMUX_PREFIX}\"|g" "${TERMUX_PYTHON_CROSSENV_BUILDHOME}/${sysconfig_module##*/}"
+		if [[ ! -f "${TERMUX_PYTHON_CROSSENV_BUILDHOME}/${sysconfig_module##*/}" ]]; then
+			cp -r "${sysconfig_module}" "${TERMUX_PYTHON_CROSSENV_BUILDHOME}"
+			sed -i "s|os.path.normpath(sys.*prefix)|\"${TERMUX_PREFIX}\"|g" "${TERMUX_PYTHON_CROSSENV_BUILDHOME}/${sysconfig_module##*/}"
+		fi
 	fi
 }
